% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeckControls.R
\name{as.tags.rdeckControls}
\alias{as.tags.rdeckControls}
\title{convert rdeckContols to html allowing !important}
\usage{
as.tags.rdeckControls(x, ...)
}
\arguments{
\item{x}{an rdeckControls widget}

\item{...}{other args passed to htmlwidgets:::toHTML}
}
\value{
html for the widget
}
\description{
convert rdeckContols to html allowing !important
}
