% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeckControls.R
\name{rdeckControls}
\alias{rdeckControls}
\title{Generic rdeck control widget}
\usage{
rdeckControls(
  targetRdeckId,
  controlType = "dropdown",
  controlData = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
A generic control widget constructor.
}
