% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeckControls.R
\name{rdeck_layer_dropdown}
\alias{rdeck_layer_dropdown}
\title{Dropdown layer selector for rdeck maps}
\usage{
rdeck_layer_dropdown(
  rdeck,
  layer_names,
  layer_group_names,
  initial_selection = NULL
)
}
\arguments{
\item{rdeck}{DESCRIPTION.}

\item{layer_names}{DESCRIPTION.}

\item{layer_group_names}{DESCRIPTION.}

\item{initial_selection}{DESCRIPTION.}
}
\value{
a htmlwidget
}
\description{
Create an external layer control for an rdeck map in rmarkdown or shiny.
}
